bug: chalk 버그 수정
